% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-rknuth.Rmd.
\name{add_seven_and_double}
\alias{add_seven_and_double}
\title{Another function defined with references to chunks.}
\usage{
add_seven_and_double(x)
}
\arguments{
\item{x}{a number}
}
\description{
This function does some math.
}